feat(frontend): wire up ESLint flat config as lint gate
eslint was installed but had no config and ran nowhere. Add flat config
(typescript-eslint, react-hooks, react-refresh), run it in CI and the
pre-commit hook, and fix all error-level findings:

- move HIDDEN_LABEL_KEYS/labelColorStyle to lib/alertUtils.ts and
  formatDuration to lib/utils.ts (react-refresh/only-export-components)
- drop unused vars/imports/params in e2e specs
- react-hooks/purity and set-state-in-effect stay warn-level as adoption
  backlog (9 findings in 6 components); react rules and no-explicit-any
  are off for Playwright code in e2e/
