+++ .githooks/pre-commit
@@ -21,6 +21,8 @@ if git diff --cached --name-only | grep -q "^frontend/"; then
   if podman exec "$FRONTEND_CONTAINER" true 2>/dev/null; then
     echo "  [Frontend] Running pnpm audit..."
     podman exec -e CI=true "$FRONTEND_CONTAINER" sh -c "cd /app && pnpm audit --audit-level=high"
+    echo "  [Frontend] Running eslint..."
+    podman exec -e CI=true "$FRONTEND_CONTAINER" sh -c "cd /app && pnpm lint"
     echo "  [Frontend] Running jscpd (code duplication)..."
     podman exec -e CI=true "$FRONTEND_CONTAINER" sh -c "cd /app && pnpm duplication"
   else
