+++ .github/workflows/ci.yml
@@ -143,6 +143,9 @@ jobs:
       - name: Audit dependencies
         run: pnpm audit --audit-level=high
 
+      - name: Lint
+        run: pnpm lint
+
       - name: Build
         run: pnpm build
