+++ frontend/package.json
@@ -9,7 +9,7 @@
     "preview": "vite preview",
     "test": "playwright test --config playwright.e2e.config.ts",
     "test:e2e": "playwright test --config playwright.e2e.config.ts",
-    "lint": "eslint src --ext ts,tsx",
+    "lint": "eslint src e2e",
     "duplication": "jscpd"
   },
   "dependencies": {
@@ -33,9 +33,13 @@
     "@types/react-dom": "^19.1.3",
     "@vitejs/plugin-react": "^6.0.3",
     "eslint": "^10.6.0",
+    "eslint-plugin-react-hooks": "^7.1.1",
+    "eslint-plugin-react-refresh": "^0.5.3",
+    "globals": "^17.7.0",
+    "jscpd": "^5.0.11",
     "tailwindcss": "^4.3.2",
     "typescript": "^6.0.3",
-    "vite": "^8.1.3",
-    "jscpd": "^5.0.11"
+    "typescript-eslint": "^8.62.1",
+    "vite": "^8.1.3"
   }
 }
