+++ frontend/src/components/alerts/AlertCard.tsx
@@ -4,7 +4,8 @@ import { cn } from '@/lib/utils'
 import { getFilterableLabels, getSilenceState, getExpiredSilence, formatSilenceDuration, tzAbbr } from '@/lib/alertUtils'
 import { renderTextWithLinks } from '@/lib/linkUtils'
 import { AlertBadge } from './AlertBadge'
-import { HIDDEN_LABEL_KEYS, LabelChip } from './LabelChip'
+import { LabelChip } from './LabelChip'
+import { HIDDEN_LABEL_KEYS } from '@/lib/alertUtils'
 import { useAlertStats } from '@/hooks/useAlerts'
 import { useFormatTime } from '@/hooks/useFormatTime'
 import { useSettingsStore } from '@/store/useSettingsStore'
