+++ frontend/src/components/alerts/AlertDetailPanel.tsx
@@ -7,7 +7,7 @@ import { TruncatableChip } from '@/components/ui/truncatable-chip'
 import { cn } from '@/lib/utils'
 import { Sheet } from '@/components/ui/sheet'
 import { AlertBadge, StatusBadge } from './AlertBadge'
-import { labelColorStyle } from './LabelChip'
+import { labelColorStyle } from '@/lib/alertUtils'
 import { AlertComments } from './AlertComments'
 import { AlertDetailHistorySection } from './AlertDetailHistorySection'
 import { AlertDetailSection } from './AlertDetailSection'
