+++ frontend/src/components/alerts/AlertListRow.tsx
@@ -1,7 +1,8 @@
 import { Bell, BellMinus, BellOff, RefreshCw, User, X } from 'lucide-react'
 import { useState, useRef, useEffect } from 'react'
 import { AlertBadge, StatusBadge } from './AlertBadge'
-import { HIDDEN_LABEL_KEYS, LabelChip } from './LabelChip'
+import { LabelChip } from './LabelChip'
+import { HIDDEN_LABEL_KEYS } from '@/lib/alertUtils'
 import { useAlertStats } from '@/hooks/useAlerts'
 import { useClaimController } from '@/hooks/useAlertClaim'
 import { useLoginGuard } from '@/hooks/useLoginGuard'
