+++ frontend/src/components/alerts/AlertListView.tsx
@@ -4,7 +4,8 @@ import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query'
 import { AlertListRow } from './AlertListRow'
 import { EmptyState } from './EmptyState'
 import { StatusBadge } from './AlertBadge'
-import { HIDDEN_LABEL_KEYS, LabelChip } from './LabelChip'
+import { LabelChip } from './LabelChip'
+import { HIDDEN_LABEL_KEYS } from '@/lib/alertUtils'
 import { Sheet } from '@/components/ui/sheet'
 import { SilenceForm } from '@/components/silences/SilenceForm'
 import { SilenceExpireModal } from '@/components/silences/SilenceExpireModal'
