+++ frontend/src/components/alerts/LabelChip.tsx
@@ -1,29 +1,10 @@
-import React, { useRef, useState } from 'react'
+import { useRef, useState } from 'react'
 import { useUIStore } from '@/store/uiStore'
 import { useSettingsStore } from '@/store/useSettingsStore'
+import { labelColorStyle } from '@/lib/alertUtils'
 import type { LabelMatcherOperator } from '@/types'
 
-export const HIDDEN_LABEL_KEYS = new Set(['alertname', 'severity', 'receiver', '@receiver'])
-
 const OPERATORS: LabelMatcherOperator[] = ['=', '!=', '=~', '!~']
-
-export function labelColorStyle(key: string, theme: 'dark' | 'light' = 'dark'): React.CSSProperties {
-  let h = 5381
-  for (let i = 0; i < key.length; i++) h = ((h << 5) + h + key.charCodeAt(i)) >>> 0
-  const hue = h % 360
-  if (theme === 'light') {
-    return {
-      backgroundColor: `hsl(${hue} 50% 90%)`,
-      color: `hsl(${hue} 70% 28%)`,
-      borderColor: `hsl(${hue} 40% 70%)`,
-    }
-  }
-  return {
-    backgroundColor: `hsl(${hue} 40% 16%)`,
-    color: `hsl(${hue} 70% 72%)`,
-    borderColor: `hsl(${hue} 35% 30%)`,
-  }
-}
 
 export function LabelChip({ labelKey, value }: { labelKey: string; value: string }) {
   const [open, setOpen] = useState(false)
