+++ frontend/src/components/silences/SilenceExpiry.tsx
@@ -1,6 +1,6 @@
 import { format } from 'date-fns'
 import { enUS } from 'date-fns/locale'
-import { cn } from '@/lib/utils'
+import { cn, formatDuration } from '@/lib/utils'
 import { tzAbbr } from '@/lib/alertUtils'
 import type { Silence } from '@/types'
 
@@ -10,22 +10,6 @@ interface SilenceExpiryProps {
 }
 
 const DATE_FMT = 'MMM d, yyyy HH:mm'
-
-/** Formats a duration in ms as e.g. "2h and 11m", "3d and 4h", "5m". */
-export function formatDuration(ms: number): string {
-  const totalMinutes = Math.max(0, Math.round(ms / 60_000))
-  const days = Math.floor(totalMinutes / (24 * 60))
-  const hours = Math.floor((totalMinutes % (24 * 60)) / 60)
-  const minutes = totalMinutes % 60
-
-  if (days > 0) {
-    return hours > 0 ? `${days}d and ${hours}h` : `${days}d`
-  }
-  if (hours > 0) {
-    return minutes > 0 ? `${hours}h and ${minutes}m` : `${hours}h`
-  }
-  return `${minutes}m`
-}
 
 function ExactDate({ value }: { value: string }) {
   return (
