+++ frontend/src/components/silences/SilenceGroupCard.tsx
@@ -2,7 +2,7 @@ import { BellMinus, Layers3, Loader2, RotateCcw } from 'lucide-react'
 import { Card, CardContent, CardHeader } from '@/components/ui/card'
 import { Button } from '@/components/ui/button'
 import { SilenceExpiry } from './SilenceExpiry'
-import { labelColorStyle } from '@/components/alerts/LabelChip'
+import { labelColorStyle } from '@/lib/alertUtils'
 import { TruncatableChip } from '@/components/ui/truncatable-chip'
 import { useSettingsStore } from '@/store/useSettingsStore'
 import type { EnrichedAlert, Silence } from '@/types'
